% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_json_data}
\alias{get_json_data}
\title{get_json_data}
\usage{
get_json_data(response = get_api_response())
}
\arguments{
\item{response}{}
}
\value{
github api response content converted to a character vector
}
\description{
get_json_data
}
\examples{
get_json_data()
}
