% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{github_api}
\alias{github_api}
\title{github_api}
\usage{
github_api(path, token = github_pat(),
  user_agent = httr::user_agent("https://github.com/adam-gruer/ozunConfIssues"))
}
\arguments{
\item{path}{the path to a github API v3 endpoint , excluding the root https://api.github.com}

\item{token}{a github personal access token}
}
\value{
a list of class github_api containing 3 elements:
        content - parsed json data,
        path - the url of the api endpoint the data was requested from,
        response - the complete response from the request
}
\description{
github_api
}
\examples{
github_api("/repos/ropensci/ozunconf17/issues")
}
