% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{github_current_user}
\alias{github_current_user}
\title{github_current_user}
\usage{
github_current_user()
}
\value{
a list of class github_api retrieving the github details for the current authenticated user
}
\description{
github_current_user
}
\examples{
github_current_user()
}
