% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{github_pat}
\alias{github_pat}
\title{github_pat}
\usage{
github_pat()
}
\value{
github personal access token from an environment variable labelled GITHUB_PAT
}
\description{
github_pat
}
\examples{
github_pat()
}
