% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{issues_df2}
\alias{issues_df2}
\title{issues_df2}
\usage{
issues_df2(content = pages_df()$content)
}
\arguments{
\item{content}{}
}
\value{
a tibble of gitgub issues for the specified repo
}
\description{
issues_df2
}
\examples{
issues_df2()
}
