% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{last_page}
\alias{last_page}
\title{last_page}
\usage{
last_page(response)
}
\arguments{
\item{response}{}
}
\value{
page number, a character vector of length 1 if the api response has a link to a last page
}
\description{
last_page
}
\examples{
last_page(github_api("repos/ropensci/ozunconf17/issues"))
}
