% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{next_page}
\alias{next_page}
\title{next_page}
\usage{
next_page(response)
}
\arguments{
\item{response}{}
}
\value{
page number, a character vector of length 1 if the api response has a link to a next page
}
\description{
next_page
}
\examples{
next_page(github_api("repos/ropensci/ozunconf17/issues"))
}
