% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{pages_df}
\alias{pages_df}
\title{pages_df}
\usage{
pages_df(pages = get_all())
}
\arguments{
\item{pages}{}
}
\value{
a tibble of of reposnes from github API
}
\description{
pages_df
}
\examples{
pages_df()
}
