% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{parse_json}
\alias{parse_json}
\title{parse_json}
\usage{
parse_json(jsondata = get_json_data())
}
\arguments{
\item{jsondata}{}
}
\value{
a list of parsed json data
}
\description{
parse_json
}
\examples{
parse_json()
}
