% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{print.github_api}
\alias{print.github_api}
\title{print.github_api}
\usage{
\method{print}{github_api}(x, ...)
}
\arguments{
\item{x}{a list of class github_api}

\item{...}{other arguments passed on to methods}
}
\value{
invisbly returns x
}
\description{
print.github_api
}
\examples{
print( github_api("repos/ropensci/ozunconf17/issues"))
}
