% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_api_response}
\alias{get_api_response}
\title{get_api_response}
\usage{

  get_api_response(url = "https://api.github.com/repos/ropensci/ozunconf17/issues",
  ...)
}
\arguments{
\item{...}{}
}
\description{
get_api_response
}
